Updated Settings script
Settings script has updated to receive button change inputs. Should be easy to integrate. There is also now a volume slider that will edit the master volume bus. Keybindings will be saved across an instance of the game, but if you quit and return to the game they will be lost. If you are implementing the inputmap, please use "Up", "Down", "Left", and "Right".